chore!: reverse middleware order
The first-added middleware is now the closest to the backend.
This is more intuitive as Announce can now be visualised as
a backend with middleware layered on top of it. It also makes
with() more useful since the most common use case of with()
is adding high-level middleware for certain handlers.